A bit more secure account page

+++ itemRecieved.php
@@ -2,19 +2,31 @@
   include_once "db_connection.php";
   session_start();
 
-  if(isset($_GET['itemID']) && isset($_GET['matchedID']))
+  if(isset($_GET['itemID']) && isset($_GET['matchedID']) && isset($_SESSION['loggedIn']) && $_SESSION['loggedIn'] == 1)
   {
 
-    $itemID = $_GET['itemID'];
-    $matchedID = $_GET['matchedID'];
+    $userID = $_SESSION["id"];
+    $itemID = test_input($_GET['itemID']);
+    $matchedID = test_input($_GET['matchedID']);
 
-    $checkQuerry = "SELECT MislayerID FROM Matched WHERE "
-
-    $sqlItem = "DELETE FROM Items WHERE ID = $itemID";
-    $conn->query($sqlItem);
-    $sqlMatched = "DELETE FROM Matched WHERE ID = $matchedID";
-    $conn->query($sqlMatched);
-    header("Location: account.php");
+    $checkQuerry = "SELECT * FROM Matched WHERE MislayerID='$userID' AND ItemID='$itemID'";
+    $checkResult = $conn->query($checkQuerry);
+    
+    if($checkResult->num_rows > 0)
+    {
+      $sqlItem = "DELETE FROM Items WHERE ID = $itemID";
+      $conn->query($sqlItem);
+      $sqlMatched = "DELETE FROM Matched WHERE ItemID = $itemID";
+      $conn->query($sqlMatched);
+      header("Location: account.php");
+    }
   }
+
+  function test_input($data) {
+  $data = trim($data);
+  $data = stripslashes($data);
+  $data = htmlspecialchars($data);
+  return $data;
+}
 ?>
